fix: cast discrepancy between 32-bit shim and 64-bit renderer, check if meshes empty before dispatch rays

+++ shared/gl_commands.h
@@ -190,7 +190,7 @@ struct GLRemixDrawArraysCommand
 {
     UINT32 mode;
     UINT32 first;
-    SIZE_T count;
+    UINT32 count;
     UINT32 enabled;  // amount of currently enabled client array kinds (<= 6)
     GLRemixClientArrayHeader headers[static_cast<UINT32>(GLRemixClientArrayType::_COUNT)];
 };
